+++ src/gui/CraftingPage.lua
@@ -168,16 +168,18 @@ function Self:UpdateSort(refresh)
         self.frame.RecipeList.NoResultsText:SetShown(false)
 
         self.sortJob = Promise:Async(function ()
-            -- Util:DebugProfileStart("Job > Start")
+            Util:DebugProfileStart("Job")
 
             Promise:GetCurrent():SetPriority(5)
 
+            Util:DebugProfileLevel("Recipe")
+
             for i,recipeID in ipairs(recipeIDs) do
-                Util:DebugProfileSegment("Recipe > GetRecipeSchematic")
+                Util:DebugProfileSegment("GetRecipeSchematic")
 
                 local recipe = C_TradeSkillUI.GetRecipeSchematic(recipeID, false)
 
-                Util:DebugProfileSegment("Recipe > Cache")
+                Util:DebugProfileSegment("GetRecipeAllocation")
 
                 local key, time = cache:Key(recipe), Prices:GetRecipeScanTime(recipe)
                 if not cache:Has(key) or cache:Get(key)[1] ~= time then
@@ -186,9 +188,11 @@ function Self:UpdateSort(refresh)
 
                 local operation = cache:Get(key)[2]
                 if operation then
+                    Util:DebugProfileSegment("GetOperationValue")
+
                     local value = Optimization:GetOperationValue(operation, method)
 
-                    Util:DebugProfileSegment("Recipe > Data Provider Insert")
+                    Util:DebugProfileSegment("DataProvider Insert")
 
                     if value and abs(value) ~= math.huge then
                         self.dataProvider:Insert({
@@ -204,16 +208,18 @@ function Self:UpdateSort(refresh)
                 Promise:YieldProgress(i, n)
             end
 
-            Util:DebugProfileSegment("Job > NoResultsText")
+            Util:DebugProfileLevelStop()
+
+            Util:DebugProfileSegment("NoResultsText")
 
             self.frame.RecipeList.NoResultsText:SetShown(self.dataProvider:IsEmpty())
 
-            Util:DebugProfileSegment("Job > Data Provider Invalidate")
+            Util:DebugProfileSegment("DataProvider Invalidate")
 
             -- Fix last recipe not getting sorted correctly
             self.dataProvider:Invalidate()
 
-            -- Util:DebugProfileStop()
+            Util:DebugProfileStop()
         end):Start(function ()
             self.progressBar:Start(n)
             return function () self.progressBar:Progress(n, n) end